\name{summary.WinRatio}
\alias{summary.WinRatio}
\title{Summarizing Win Ratio Results}
\usage{
\method{summary}{WinRatio}(object, \dots, digits = 2)
}
\arguments{
\item{object}{an object of class \code{'WinRatio'}.}
\item{\dots}{additional arguments affecting the summary produced.}
\item{digits}{the number of significant digits to use when printing win ratio and 95\% confidence interval.}
}
\description{
This function is used for summarizing the results for class \code{'WinRatio'} objects.
}

\examples{
## For examples see example(winratio)
}
