\name{WinRatio-package}
\alias{WinRatio-package}
\docType{package}
\title{
  Win Ratio for Prioritized Outcomes and 95\% Confidence Interval
}
\description{
  Calculate the win ratio for prioritized outcomes and the 95\% confidence interval based on Bebu and Lachin (2016). Three type of outcomes can be analyzed: survival "failure-time" events, repeated survival "failure-time" events and continuous or ordinal "non-failure time" events that are captured at specific time-points in the study.
}
\author{
Kevin Duarte, 
  Joao Pedro Ferreira

Maintainer: Kevin Duarte <k.duarte@chru-nancy.fr>
}
\references{
  Pocock S.J., Ariti C.A., Collier T.J., Wang D. The win ratio: a new approach to the analysis of composite endpoints in clinical trials based on clinical priorities. \emph{Eur Heart Journal}. 2012; \bold{33}:176-182.
  DOI: \href{https://doi.org/10.1093/eurheartj/ehr352}{10.1093/eurheartj/ehr352}.

  Bebu I. and Lachin J.M. Large sample inference for a win ratio analysis of a composite outcome based on prioritized components. \emph{Biostatistics}. 2016; \bold{17}:178-187.
   DOI: \href{https://doi.org/10.1093/biostatistics/kxv032}{10.1093/biostatistics/kxv032}.
}
